backpy/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
email-validator==2.1.0
Werkzeug==2.3.7
python-dotenv==1.0.0
passlib==1.7.4
PyJWT==2.8.0
flask-cors==4.0.0 